+++ simpegEM1D/__init__.py
@@ -1,7 +1,7 @@
 from .simulation import BaseEM1DSimulation, EM1DFMSimulation, EM1DTMSimulation
 from .survey import BaseEM1DSurvey, EM1DSurveyFD, EM1DSurveyTD
 from .DigFilter import *
-from .EM1DAnalytics import *
+from .analytics import *
 from .RTEfun import rTEfunfwd, rTEfunjac
 from .Waveform import *
 from .Waveforms import (
