+++ simpegEM1D/analytics.py
