+++ simpegEM1D/simulation.py
@@ -2,8 +2,8 @@ from SimPEG import maps, utils, props
 from SimPEG.simulation import BaseSimulation
 import numpy as np
 from .survey import BaseEM1DSurvey
+from .kernels import *
 from scipy.constants import mu_0
-from .RTEfun_vec import rTEfunfwd, rTEfunjac
 from scipy.interpolate import InterpolatedUnivariateSpline as iuSpline
 import properties
 
@@ -11,10 +11,7 @@ from empymod import filters
 from empymod.transform import dlf, fourier_dlf, get_dlf_points
 from empymod.utils import check_hankel
 
-try:
-    from simpegEM1D.m_rTE_Fortran import rte_fortran
-except ImportError as e:
-    rte_fortran = None
+
 
 
 class BaseEM1DSimulation(BaseSimulation):
@@ -93,196 +90,6 @@ class BaseEM1DSimulation(BaseSimulation):
 
         # if self.hankel_pts_per_dec != 0:
         #     raise NotImplementedError()
-
-    def hz_kernel_vertical_magnetic_dipole(
-        self, lamda, f, n_layer, sig, chi, depth, h, z,
-        flag, I, output_type='response'
-    ):
-
-        """
-            Kernel for vertical magnetic component (Hz) due to
-            vertical magnetic diopole (VMD) source in (kx,ky) domain
-
-        """
-        u0 = lamda
-        coefficient_wavenumber = 1/(4*np.pi)*lamda**3/u0
-
-        n_frequency = self.survey.n_frequency
-        n_layer = self.survey.n_layer
-        n_filter = self.n_filter
-
-        if output_type == 'sensitivity_sigma':
-            drTE = np.zeros(
-                [n_layer, n_frequency, n_filter],
-                dtype=np.complex128, order='F'
-            )
-            if rte_fortran is None:
-                drTE = rTEfunjac(
-                    n_layer, f, lamda, sig, chi, depth, self.survey.half_switch
-                )
-            else:
-                rte_fortran.rte_sensitivity(
-                    f, lamda, sig, chi, depth, self.survey.half_switch, drTE,
-                    n_layer, n_frequency, n_filter
-                    )
-
-            kernel = drTE * np.exp(-u0*(z+h)) * coefficient_wavenumber
-        else:
-            rTE = np.empty(
-                [n_frequency, n_filter], dtype=np.complex128, order='F'
-            )
-            if rte_fortran is None:
-                    rTE = rTEfunfwd(
-                        n_layer, f, lamda, sig, chi, depth,
-                        self.survey.half_switch
-                    )
-            else:
-                rte_fortran.rte_forward(
-                    f, lamda, sig, chi, depth, self.survey.half_switch,
-                    rTE, n_layer, n_frequency, n_filter
-                )
-
-            kernel = rTE * np.exp(-u0*(z+h)) * coefficient_wavenumber
-            if output_type == 'sensitivity_height':
-                kernel *= -2*u0
-
-        return kernel * I
-
-        # Note
-        # Here only computes secondary field.
-        # I am not sure why it does not work if we add primary term.
-        # This term can be analytically evaluated, where h = 0.
-        #     kernel = (
-        #         1./(4*np.pi) *
-        #         (np.exp(u0*(z-h))+rTE * np.exp(-u0*(z+h)))*lamda**3/u0
-        #     )
-
-    # TODO: make this to take a vector rather than a single frequency
-    def hz_kernel_circular_loop(
-        self, lamda, f, n_layer, sig, chi, depth, h, z, I, a,
-        flag,  output_type='response'
-    ):
-
-        """
-
-        Kernel for vertical magnetic component (Hz) at the center
-        due to circular loop source in (kx,ky) domain
-
-        .. math::
-
-            H_z = \\frac{Ia}{2} \int_0^{\infty} [e^{-u_0|z+h|} +
-            \\r_{TE}e^{u_0|z-h|}]
-            \\frac{\lambda^2}{u_0} J_1(\lambda a)] d \lambda
-
-        """
-
-        n_frequency = self.survey.n_frequency
-        n_layer = self.survey.n_layer
-        n_filter = self.n_filter
-
-        w = 2*np.pi*f
-        u0 = lamda
-        radius = np.empty([n_frequency, n_filter], order='F')
-        radius[:, :] = np.tile(a.reshape([-1, 1]), (1, n_filter))
-
-        coefficient_wavenumber = I*radius*0.5*lamda**2/u0
-
-        if output_type == 'sensitivity_sigma':
-            drTE = np.empty(
-                [n_layer, n_frequency, n_filter],
-                dtype=np.complex128, order='F'
-            )
-            if rte_fortran is None:
-                    drTE[:, :] = rTEfunjac(
-                        n_layer, f, lamda, sig, chi, depth,
-                        self.survey.half_switch
-                    )
-            else:
-                rte_fortran.rte_sensitivity(
-                    f, lamda, sig, chi, depth, self.survey.half_switch,
-                    drTE, n_layer, n_frequency, n_filter
-                )
-
-            kernel = drTE * np.exp(-u0*(z+h)) * coefficient_wavenumber
-        else:
-            rTE = np.empty(
-                [n_frequency, n_filter], dtype=np.complex128, order='F'
-            )
-            if rte_fortran is None:
-                rTE[:, :] = rTEfunfwd(
-                    n_layer, f, lamda, sig, chi, depth, self.survey.half_switch
-                )
-            else:
-                rte_fortran.rte_forward(
-                    f, lamda, sig, chi, depth, self.survey.half_switch,
-                    rTE, n_layer, n_frequency, n_filter
-                )
-
-            if flag == 'secondary':
-                kernel = rTE * np.exp(-u0*(z+h)) * coefficient_wavenumber
-            else:
-                kernel = rTE * (
-                    np.exp(-u0*(z+h)) + np.exp(u0*(z-h))
-                ) * coefficient_wavenumber
-
-            if output_type == 'sensitivity_height':
-                kernel *= -2*u0
-
-        return kernel
-
-    def hz_kernel_horizontal_electric_dipole(
-        self, lamda, f, n_layer, sig, chi, depth, h, z,
-        flag, output_type='response'
-    ):
-
-        """
-            Kernel for vertical magnetic field (Hz) due to
-            horizontal electric diopole (HED) source in (kx,ky) domain
-
-        """
-        n_frequency = self.survey.n_frequency
-        n_layer = self.survey.n_layer
-        n_filter = self.n_filter
-
-        u0 = lamda
-        coefficient_wavenumber = 1/(4*np.pi)*lamda**2/u0
-
-        if output_type == 'sensitivity_sigma':
-            drTE = np.zeros(
-                [n_layer, n_frequency, n_filter], dtype=np.complex128,
-                order='F'
-            )
-            if rte_fortran is None:
-                drTE = rTEfunjac(
-                    n_layer, f, lamda, sig, chi, depth, self.survey.half_switch
-                )
-            else:
-                rte_fortran.rte_sensitivity(
-                    f, lamda, sig, chi, depth, self.survey.half_switch,
-                    drTE, n_layer, n_frequency, n_filter
-                )
-
-            kernel = drTE * np.exp(-u0*(z+h)) * coefficient_wavenumber
-        else:
-            rTE = np.empty(
-                [n_frequency, n_filter], dtype=np.complex128, order='F'
-            )
-            if rte_fortran is None:
-                rTE = rTEfunfwd(
-                        n_layer, f, lamda, sig, chi, depth,
-                        self.survey.half_switch
-                )
-            else:
-                rte_fortran.rte_forward(
-                    f, lamda, sig, chi, depth, self.survey.half_switch,
-                    rTE, n_layer, n_frequency, n_filter
-                )
-
-            kernel = rTE * np.exp(-u0*(z+h)) * coefficient_wavenumber
-            if output_type == 'sensitivity_height':
-                kernel *= -2*u0
-
-        return kernel
 
     # make it as a property?
 
@@ -400,8 +207,8 @@ class BaseEM1DSimulation(BaseSimulation):
         if output_type == 'response':
             # for simulation
             if self.survey.src_type == 'VMD':
-                hz = self.hz_kernel_vertical_magnetic_dipole(
-                    lambd, f, n_layer,
+                hz = hz_kernel_vertical_magnetic_dipole(
+                    self, lambd, f, n_layer,
                     sig, chi, depth, h, z,
                     flag, I, output_type=output_type
                 )
@@ -411,8 +218,8 @@ class BaseEM1DSimulation(BaseSimulation):
                 PJ = (hz, None, None)  # PJ0
 
             elif self.survey.src_type == 'CircularLoop':
-                hz = self.hz_kernel_circular_loop(
-                    lambd, f, n_layer,
+                hz = hz_kernel_circular_loop(
+                    self, lambd, f, n_layer,
                     sig, chi, depth, h, z, I, r,
                     flag, output_type=output_type
                 )
@@ -424,8 +231,8 @@ class BaseEM1DSimulation(BaseSimulation):
             # TODO: This has not implemented yet!
             elif self.survey.src_type == "piecewise_line":
                 # Need to compute y
-                hz = self.hz_kernel_horizontal_electric_dipole(
-                    lambd, f, n_layer,
+                hz = hz_kernel_horizontal_electric_dipole(
+                    self, lambd, f, n_layer,
                     sig, chi, depth, h, z, I, r,
                     flag, output_type=output_type
                 )
@@ -440,8 +247,8 @@ class BaseEM1DSimulation(BaseSimulation):
 
             # for simulation
             if self.survey.src_type == 'VMD':
-                hz = self.hz_kernel_vertical_magnetic_dipole(
-                    lambd, f, n_layer,
+                hz = hz_kernel_vertical_magnetic_dipole(
+                    self, lambd, f, n_layer,
                     sig, chi, depth, h, z,
                     flag, I, output_type=output_type
                 )
@@ -450,8 +257,8 @@ class BaseEM1DSimulation(BaseSimulation):
 
             elif self.survey.src_type == 'CircularLoop':
 
-                hz = self.hz_kernel_circular_loop(
-                    lambd, f, n_layer,
+                hz = hz_kernel_circular_loop(
+                    self, lambd, f, n_layer,
                     sig, chi, depth, h, z, I, r,
                     flag, output_type=output_type
                 )
@@ -467,8 +274,8 @@ class BaseEM1DSimulation(BaseSimulation):
 
             # for simulation
             if self.survey.src_type == 'VMD':
-                hz = self.hz_kernel_vertical_magnetic_dipole(
-                    lambd, f, n_layer,
+                hz = hz_kernel_vertical_magnetic_dipole(
+                    self, lambd, f, n_layer,
                     sig, chi, depth, h, z,
                     flag, I, output_type=output_type
                 )
@@ -477,8 +284,8 @@ class BaseEM1DSimulation(BaseSimulation):
 
             elif self.survey.src_type == 'CircularLoop':
 
-                hz = self.hz_kernel_circular_loop(
-                    lambd, f, n_layer,
+                hz = hz_kernel_circular_loop(
+                    self, lambd, f, n_layer,
                     sig, chi, depth, h, z, I, r,
                     flag, output_type=output_type
                 )
